feat(ui): add harpoon marks to statusline
This commit adds harpoon marks to the statusline.
It defines new highlight groups for the active and inactive harpoon marks,
the harpoon icon, and the separator. It also adds a new module to the statusline
that displays the harpoon marks.
Finally, it removes the harpoon marks from the incline plugin.